Move setup_locale from contest to base handler
