+++ cms/server/contest/handlers/contest.py
@@ -41,20 +41,16 @@ from six import iterkeys, iteritems
 import ipaddress
 import logging
 import pickle
-
 from datetime import timedelta
 
 import tornado.web
-
 from sqlalchemy.orm import contains_eager
-from werkzeug.datastructures import LanguageAccept
-from werkzeug.http import parse_accept_header
 
 from cms import config
 from cms.db import Contest, Participation, User
 from cms.server import compute_actual_phase, file_handler_gen, \
     create_url_builder
-from cms.locale import filter_language_codes, choose_language_code
+from cms.locale import filter_language_codes
 from cmscommon.datetime import get_timezone, make_datetime, make_timestamp
 
 from .base import BaseHandler
@@ -102,7 +98,6 @@ class ContestHandler(BaseHandler):
         self.contest_url = None
 
     def prepare(self):
-        super(ContestHandler, self).prepare()
         self.choose_contest()
 
         if self.contest.allowed_localizations:
@@ -113,7 +108,7 @@ class ContestHandler(BaseHandler):
                 (k, v) for k, v in iteritems(self.available_translations)
                 if k in lang_codes)
 
-        self.setup_locale()
+        super(ContestHandler, self).prepare()
 
         if self.is_multi_contest():
             self.contest_url = \
@@ -309,34 +304,6 @@ class ContestHandler(BaseHandler):
 
         return participation
 
-    def setup_locale(self):
-        lang_codes = list(iterkeys(self.available_translations))
-
-        browser_langs = parse_accept_header(
-            self.request.headers.get("Accept-Language", ""),
-            LanguageAccept).values()
-        automatic_lang = choose_language_code(browser_langs, lang_codes)
-        if automatic_lang is None:
-            automatic_lang = lang_codes[0]
-        self.automatic_translation = \
-            self.available_translations[automatic_lang]
-
-        cookie_lang = self.get_cookie("language", None)
-        if cookie_lang is not None:
-            chosen_lang = \
-                choose_language_code([cookie_lang, automatic_lang], lang_codes)
-            if chosen_lang == cookie_lang:
-                self.cookie_translation = \
-                    self.available_translations[cookie_lang]
-        else:
-            chosen_lang = automatic_lang
-        self.translation = self.available_translations[chosen_lang]
-
-        self._ = self.translation.gettext
-        self.n_ = self.translation.ngettext
-
-        self.set_header("Content-Language", chosen_lang)
-
     @staticmethod
     def _get_token_status(obj):
         """Return the status of the tokens for the given object.
@@ -401,14 +368,6 @@ class ContestHandler(BaseHandler):
             ret["tokens_tasks"] = 2  # all infinite
         else:
             ret["tokens_tasks"] = 1  # all finite or mixed
-
-        ret["available_translations"] = self.available_translations
-
-        ret["cookie_translation"] = self.cookie_translation
-        ret["automatic_translation"] = self.automatic_translation
-
-        ret["translation"] = self.translation
-        ret["_"] = self._
 
         return ret
